rename configurator to config-loader
